requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confection==0.0.3
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
cymem==2.0.7
dante-tokenizer==0.2.0
fastapi==0.86.0
ffmpy==0.3.0
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gradio==3.9.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.10.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
linkify-it-py==1.0.3
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mdit-py-plugins==0.3.1
mdurl==0.1.2
multidict==6.0.2
murmurhash==1.0.9
nltk==3.7
numpy==1.23.4
orjson==3.8.1
packaging==21.3
pandas==1.5.1
paramiko==2.12.0
pathy==0.6.2
Pillow==9.3.0
preshed==3.0.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pydub==0.25.1
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
starlette==0.20.4
thinc==8.1.5
tokenizers==0.13.2
torch==1.13.0
tqdm==4.64.1
transformers==4.24.0
typer==0.7.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.12
uvicorn==0.19.0
wasabi==0.10.1
websockets==10.4
yarl==1.8.1
option==2.1.0